requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# job_classifier/server.py: 2
Flask == 1.1.2

# job_classifier/utils/match_rules.py: 1
PyYAML == 5.3.1

# job_classifier/utils/utils.py: 5
beautifulsoup4 == 4.9.2

# job_classifier/utils/match_rules.py: 1
cronyo == 0.4.2

# job_classifier/utils/utils.py: 7,8
nltk == 3.5

# job_classifier/server.py: 1
numpy == 1.19.2

# job_classifier/utils/utils.py: 6
scikit_learn == 0.23.2

# job_classifier/utils/match_rules.py: 1
yoc == 0.0.0
